requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
requests==2.31.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
Pillow==10.1.0
setuptools>=65.0.0
wheel>=0.37.0
mysqlclient>=2.1.0
